# Use MappingQCache for curved cells if possible to speed up particle sort