@jdannberg 2021-02-01
It looks like some tests fail because the mapping hasn't been initialized yet. 
Otherwise, this is great! Thank you!