@gassmoeller 2021-02-02
I fixed the forgotten initialization (the one right after constructing the coarse mesh, but before the initial refinement). I still have one minor change in the test `hollow_sphere_gmg_project_q1`, but it only changes the nonlinear residuum of a later iteration, and in fact reduces the nonlinear residual, so I guess we are good with that. I updated the one test result. Ready from my side.